feat(component-picker) Pull component picker into a separate file and add filters
